refactor: IndexedDB to idb
